+++ docs/source-files.js
@@ -1,3 +1,3 @@
 var N = null;var sourcesIndex = {};
-sourcesIndex["my_iot"] = {"name":"","dirs":[{"name":"services","files":["buienradar.rs","clock.rs","db.rs"]},{"name":"templates","files":["base.rs","index.rs","measurement.rs","navbar.rs","sensor.rs","services.rs"]}],"files":["db.rs","logging.rs","main.rs","measurement.rs","receiver.rs","services.rs","settings.rs","statics.rs","templates.rs","types.rs","value.rs","web.rs"]};
+sourcesIndex["my_iot"] = {"name":"","dirs":[{"name":"services","files":["buienradar.rs","clock.rs","db.rs"]},{"name":"templates","files":["base.rs","index.rs","measurement.rs","navbar.rs","sensor.rs","status.rs"]}],"files":["db.rs","logging.rs","main.rs","measurement.rs","receiver.rs","services.rs","settings.rs","statics.rs","templates.rs","types.rs","value.rs","web.rs"]};
 createSourceSidebar();