+++ src/templates.rs
@@ -5,6 +5,6 @@ pub mod index;
 pub mod measurement;
 pub mod navbar;
 pub mod sensor;
-pub mod services;
+pub mod status;
 
 const DATE_FORMAT: &str = "%b %d, %H:%M:%S";