+++ src/templates/navbar.rs
@@ -22,9 +22,9 @@ markup::define! {
                             span.icon { i.fas."fa-home" {} }
                             span { "Home" }
                         }
-                        a."navbar-item"[href = "/services"] {
-                            span.icon { i.fas."fa-cog" {} }
-                            span { "Services" }
+                        a."navbar-item"[href = "/status"] {
+                            span.icon { i.fas."fa-info" {} }
+                            span { "Status" }
                         }
                         a."navbar-item"[href = "/sensors"] {
                             span.icon { i.fas."fa-chart-line" {} }