+++ src/templates/services.rs
@@ -1,7 +1,0 @@
-//! Services overview page.
-
-markup::define! {
-    Services {
-
-    }
-}