+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## `0.7.0`
 
 - Chore: refactor service initialisation & running
+- New: add favicons
 
 ## `0.6.0`