+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # `my-iot`
 
-## `0.7.0`
+## `master`
 
 - Chore: refactor service initialisation & running
 - New: add favicons
 - Break: introduce service IDs
+- New: set thread names
 
 ## `0.6.0`