+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - New: add favicons
 - Break: introduce service IDs
 - New: set thread names
+- New: status page
 
 ## `0.6.0`